+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## 1.7.25
 * Network Security Groups: Fix bug where a SecurityRule without a source throws a meaningful exception
+* Network Security Groups: Add rule to existing security group
 
 ## 1.7.24
 * Network Interface: Adds support for network interface creation.